move metrics_util (which contains the "Installation" singleton)
