+++ lib/tests/streamlit/runtime/metrics_util_test.py
@@ -15,7 +15,7 @@
 import unittest
 from unittest.mock import patch, mock_open, MagicMock
 
-from streamlit import metrics_util
+from streamlit.runtime import metrics_util
 
 MAC = "mac"
 UUID = "uuid"
@@ -35,9 +35,9 @@ class MetricsUtilTest(unittest.TestCase):
         """Test getting the machine id from /etc"""
         file_data = "etc"
 
-        with patch("streamlit.metrics_util.uuid.getnode", return_value=MAC), patch(
-            "streamlit.metrics_util.open", mock_open(read_data=file_data), create=True
-        ), patch("streamlit.metrics_util.os.path.isfile") as path_isfile:
+        with patch("streamlit.runtime.metrics_util.uuid.getnode", return_value=MAC), patch(
+            "streamlit.runtime.metrics_util.open", mock_open(read_data=file_data), create=True
+        ), patch("streamlit.runtime.metrics_util.os.path.isfile") as path_isfile:
 
             def path_isfile(path):
                 return path == "/etc/machine-id"
@@ -49,9 +49,9 @@ class MetricsUtilTest(unittest.TestCase):
         """Test getting the machine id from /var/lib/dbus"""
         file_data = "dbus"
 
-        with patch("streamlit.metrics_util.uuid.getnode", return_value=MAC), patch(
-            "streamlit.metrics_util.open", mock_open(read_data=file_data), create=True
-        ), patch("streamlit.metrics_util.os.path.isfile") as path_isfile:
+        with patch("streamlit.runtime.metrics_util.uuid.getnode", return_value=MAC), patch(
+            "streamlit.runtime.metrics_util.open", mock_open(read_data=file_data), create=True
+        ), patch("streamlit.runtime.metrics_util.os.path.isfile") as path_isfile:
 
             def path_isfile(path):
                 return path == "/var/lib/dbus/machine-id"
@@ -62,8 +62,8 @@ class MetricsUtilTest(unittest.TestCase):
     def test_machine_id_v3_from_node(self):
         """Test getting the machine id as the mac address"""
 
-        with patch("streamlit.metrics_util.uuid.getnode", return_value=MAC), patch(
-            "streamlit.metrics_util.os.path.isfile", return_value=False
+        with patch("streamlit.runtime.metrics_util.uuid.getnode", return_value=MAC), patch(
+            "streamlit.runtime.metrics_util.os.path.isfile", return_value=False
         ):
 
             machine_id = metrics_util._get_machine_id_v3()
